fft_block: Use the stream_to_vector_overlap block instead of the classic stream_to_vector block.
For the spectral density measurement, we use an overlapping factor saved
for each window. For all the other measurements, we do not use
overlapping yet.
Also, for the spectral density we added a mechanism to use a number of
averages that should be overlapped. For now, the value will remain at its
default, 1.
The spectrum analyzer now allows the user to also change the gain mode
for each channel: high or low.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>